requirements.txt
PyPDF2==3.0.1
pydantic==2.10.6
openai==1.68.2
azure-search-documents==11.5.2
azure-core==1.32.0
langchain==0.3.21
python-dotenv==1.1.0
